[PIPE-513] Moves table_spec definitions out of commands
