+++ usaspending_api/etl/management/commands/load_query_to_delta.py
@@ -11,260 +11,7 @@ from usaspending_api.common.helpers.spark_helpers import (
     get_jvm_logger,
 )
 from usaspending_api.config import CONFIG
-from usaspending_api.recipient.delta_models import (
-    recipient_lookup_load_sql_string_list,
-    RECIPIENT_LOOKUP_POSTGRES_COLUMNS,
-    recipient_profile_create_sql_string,
-    recipient_profile_load_sql_strings,
-    RECIPIENT_PROFILE_POSTGRES_COLUMNS,
-    rpt_recipient_lookup_create_sql_string,
-    RPT_RECIPIENT_LOOKUP_DELTA_COLUMNS,
-    SAM_RECIPIENT_COLUMNS,
-    SAM_RECIPIENT_POSTGRES_COLUMNS,
-    sam_recipient_create_sql_string,
-    sam_recipient_load_sql_string,
-    RPT_RECIPIENT_PROFILE_DELTA_COLUMNS,
-)
-from usaspending_api.recipient.models import RecipientLookup, RecipientProfile
-from usaspending_api.search.delta_models.award_search import (
-    AWARD_SEARCH_COLUMNS,
-    award_search_create_sql_string,
-    award_search_incremental_load_sql_string,
-    award_search_overwrite_load_sql_string,
-    AWARD_SEARCH_POSTGRES_COLUMNS,
-    award_search_incremental_delete_temp_schema,
-)
-from usaspending_api.search.delta_models.subaward_search import (
-    SUBAWARD_SEARCH_COLUMNS,
-    subaward_search_create_sql_string,
-    subaward_search_load_sql_string,
-    SUBAWARD_SEARCH_POSTGRES_COLUMNS,
-    SUBAWARD_SEARCH_POSTGRES_VECTORS,
-)
-from usaspending_api.search.models import AwardSearch, SubawardSearch, SummaryStateView, TransactionSearch
-from usaspending_api.transactions.delta_models import (
-    transaction_search_create_sql_string,
-    transaction_search_load_sql_string,
-    transaction_current_cd_lookup_create_sql_string,
-    transaction_current_cd_lookup_load_sql_string,
-    TRANSACTION_SEARCH_POSTGRES_COLUMNS,
-    TRANSACTION_SEARCH_POSTGRES_GOLD_COLUMNS,
-    TRANSACTION_CURRENT_CD_LOOKUP_COLUMNS,
-    SUMMARY_STATE_VIEW_COLUMNS,
-    summary_state_view_create_sql_string,
-    summary_state_view_load_sql_string,
-    SUMMARY_STATE_VIEW_POSTGRES_COLUMNS,
-)
-
-TABLE_SPEC = {
-    "award_search": {
-        "model": AwardSearch,
-        "is_from_broker": False,
-        "source_query": award_search_overwrite_load_sql_string,
-        "source_query_incremental": award_search_incremental_load_sql_string,
-        "source_database": None,
-        "source_table": None,
-        "destination_database": "rpt",
-        "swap_table": "award_search",
-        "swap_schema": "rpt",
-        "unique_identifiers": ["award_id"],
-        "delta_table_create_sql": award_search_create_sql_string,
-        "source_schema": AWARD_SEARCH_POSTGRES_COLUMNS,
-        "custom_schema": "recipient_hash STRING, federal_accounts STRING, cfdas ARRAY<STRING>,"
-        " tas_components ARRAY<STRING>",
-        "column_names": list(AWARD_SEARCH_COLUMNS),
-        "postgres_seq_name": None,
-        "tsvectors": None,
-        "postgres_partition_spec": None,
-        # Incremental Updates
-        "source_query_incremental": award_search_incremental_load_sql_string,
-        "incremental_delete_temp_schema": award_search_incremental_delete_temp_schema,
-        "delta_table_load_version_key": "award_search",
-    },
-    "recipient_lookup": {
-        "model": RecipientLookup,
-        "is_from_broker": False,
-        "source_query": recipient_lookup_load_sql_string_list,
-        "source_database": None,
-        "source_table": None,
-        "destination_database": "rpt",
-        "swap_table": "recipient_lookup",
-        "swap_schema": "rpt",
-        "unique_identifiers": ["recipient_hash"],
-        "delta_table_create_sql": rpt_recipient_lookup_create_sql_string,
-        "source_schema": RECIPIENT_LOOKUP_POSTGRES_COLUMNS,
-        "custom_schema": "recipient_hash STRING",
-        "column_names": list(RPT_RECIPIENT_LOOKUP_DELTA_COLUMNS),
-        "postgres_seq_name": "recipient_lookup_id_seq",
-        "tsvectors": None,
-        "postgres_partition_spec": None,
-        # Incremental Updates
-        "source_query_incremental": None,
-        "incremental_delete_temp_schema": None,
-        "delta_table_load_version_key": None,
-    },
-    "recipient_profile": {
-        "model": RecipientProfile,
-        "is_from_broker": False,
-        "source_query": recipient_profile_load_sql_strings,
-        "source_database": None,
-        "source_table": None,
-        "destination_database": "rpt",
-        "swap_table": "recipient_profile",
-        "swap_schema": "rpt",
-        "unique_identifiers": ["recipient_hash", "recipient_level"],
-        "delta_table_create_sql": recipient_profile_create_sql_string,
-        "source_schema": RECIPIENT_PROFILE_POSTGRES_COLUMNS,
-        "custom_schema": "recipient_hash STRING",
-        "column_names": list(RPT_RECIPIENT_PROFILE_DELTA_COLUMNS),
-        "postgres_seq_name": "recipient_profile_id_seq",
-        "tsvectors": None,
-        "postgres_partition_spec": None,
-        # Incremental Updates
-        "source_query_incremental": None,
-        "incremental_delete_temp_schema": None,
-        "delta_table_load_version_key": None,
-    },
-    "summary_state_view": {
-        "model": SummaryStateView,
-        "is_from_broker": False,
-        "source_query": summary_state_view_load_sql_string,
-        "source_database": None,
-        "source_table": None,
-        "destination_database": "rpt",
-        "swap_table": "summary_state_view",
-        "swap_schema": "rpt",
-        "unique_identifiers": ["duh"],
-        "delta_table_create_sql": summary_state_view_create_sql_string,
-        "source_schema": SUMMARY_STATE_VIEW_POSTGRES_COLUMNS,
-        "custom_schema": "duh STRING",
-        "column_names": list(SUMMARY_STATE_VIEW_COLUMNS),
-        "postgres_seq_name": None,
-        "tsvectors": None,
-        "postgres_partition_spec": None,
-        # Incremental Updates
-        "source_query_incremental": None,
-        "incremental_delete_temp_schema": None,
-        "delta_table_load_version_key": None,
-    },
-    "sam_recipient": {
-        "model": None,
-        "is_from_broker": True,
-        "source_query": sam_recipient_load_sql_string,
-        "source_database": None,
-        "source_table": None,
-        "destination_database": "int",
-        "swap_table": "duns",
-        "swap_schema": "int",
-        "unique_identifiers": ["broker_duns_id"],
-        "delta_table_create_sql": sam_recipient_create_sql_string,
-        "source_schema": SAM_RECIPIENT_POSTGRES_COLUMNS,
-        "custom_schema": None,
-        "column_names": list(SAM_RECIPIENT_COLUMNS),
-        "postgres_seq_name": None,
-        "tsvectors": None,
-        "postgres_partition_spec": None,
-        # Incremental Updates
-        "source_query_incremental": None,
-        "incremental_delete_temp_schema": None,
-        "delta_table_load_version_key": None,
-    },
-    "transaction_search": {
-        "model": TransactionSearch,
-        "is_from_broker": False,
-        "source_query": transaction_search_load_sql_string,
-        "source_database": None,
-        "source_table": None,
-        "destination_database": "rpt",
-        "swap_table": "transaction_search",
-        "swap_schema": "rpt",
-        "unique_identifiers": ["transaction_id"],
-        "delta_table_create_sql": transaction_search_create_sql_string,
-        "source_schema": TRANSACTION_SEARCH_POSTGRES_COLUMNS,
-        "custom_schema": "recipient_hash STRING, federal_accounts STRING, parent_recipient_hash STRING",
-        "column_names": list(TRANSACTION_SEARCH_POSTGRES_COLUMNS),
-        "postgres_seq_name": None,
-        "tsvectors": None,
-        "postgres_partition_spec": None,
-        # Incremental Updates
-        "source_query_incremental": None,
-        "incremental_delete_temp_schema": None,
-        "delta_table_load_version_key": None,
-    },
-    "transaction_search_gold": {
-        "model": TransactionSearch,
-        "is_from_broker": False,
-        "source_query": transaction_search_load_sql_string,
-        "source_database": None,
-        "source_table": None,
-        "destination_database": "rpt",
-        "swap_table": "transaction_search",
-        "swap_schema": "rpt",
-        "unique_identifiers": ["transaction_id"],
-        "delta_table_create_sql": transaction_search_create_sql_string,
-        "source_schema": TRANSACTION_SEARCH_POSTGRES_GOLD_COLUMNS,
-        "custom_schema": "recipient_hash STRING, federal_accounts STRING, parent_recipient_hash STRING",
-        "column_names": list(TRANSACTION_SEARCH_POSTGRES_GOLD_COLUMNS),
-        "postgres_seq_name": None,
-        "tsvectors": None,
-        "postgres_partition_spec": {
-            "partition_keys": ["is_fpds"],
-            "partitioning_form": "LIST",
-            "partitions": [
-                {"table_suffix": "_fpds", "partitioning_clause": "FOR VALUES IN (TRUE)"},
-                {"table_suffix": "_fabs", "partitioning_clause": "FOR VALUES IN (FALSE)"},
-            ],
-        },
-        # Incremental Updates
-        "source_query_incremental": None,
-        "incremental_delete_temp_schema": None,
-        "delta_table_load_version_key": None,
-    },
-    "transaction_current_cd_lookup": {
-        "model": None,
-        "is_from_broker": False,
-        "source_query": transaction_current_cd_lookup_load_sql_string,
-        "source_database": None,
-        "source_table": None,
-        "destination_database": "int",
-        "swap_table": "transaction_current_cd_lookup",
-        "swap_schema": "int",
-        "unique_identifiers": ["transaction_id"],
-        "delta_table_create_sql": transaction_current_cd_lookup_create_sql_string,
-        "source_schema": TRANSACTION_CURRENT_CD_LOOKUP_COLUMNS,
-        "custom_schema": "",
-        "column_names": list(TRANSACTION_CURRENT_CD_LOOKUP_COLUMNS),
-        "postgres_seq_name": None,
-        "tsvectors": None,
-        "postgres_partition_spec": None,
-        # Incremental Updates
-        "source_query_incremental": None,
-        "incremental_delete_temp_schema": None,
-        "delta_table_load_version_key": None,
-    },
-    "subaward_search": {
-        "model": SubawardSearch,
-        "is_from_broker": False,
-        "source_query": subaward_search_load_sql_string,
-        "source_database": None,
-        "source_table": None,
-        "destination_database": "rpt",
-        "swap_table": "subaward_search",
-        "swap_schema": "rpt",
-        "unique_identifiers": ["broker_subaward_id"],
-        "delta_table_create_sql": subaward_search_create_sql_string,
-        "source_schema": SUBAWARD_SEARCH_POSTGRES_COLUMNS,
-        "custom_schema": "treasury_account_identifiers ARRAY<INTEGER>",
-        "column_names": list(SUBAWARD_SEARCH_COLUMNS),
-        "postgres_seq_name": None,
-        "tsvectors": SUBAWARD_SEARCH_POSTGRES_VECTORS,
-        "postgres_partition_spec": None,
-        # Incremental Updates
-        "source_query_incremental": None,
-        "incremental_delete_temp_schema": None,
-        "delta_table_load_version_key": None,
-    },
-}
+from usaspending_api.etl.management.helpers.table_specifications import DATABRICKS_GENERATED_TABLE_SPEC as TABLE_SPEC
 
 
 class Command(BaseCommand):
