+++ usaspending_api/etl/management/commands/load_table_to_delta.py
@@ -1,15 +1,5 @@
 from django.core.management import BaseCommand
 from pyspark.sql import SparkSession
-
-from usaspending_api.awards.delta_models import (
-    AWARDS_COLUMNS,
-    awards_sql_string,
-    FINANCIAL_ACCOUNTS_BY_AWARDS_COLUMNS,
-    financial_accounts_by_awards_sql_string,
-    BROKER_SUBAWARDS_COLUMNS,
-    broker_subawards_sql_string,
-)
-from usaspending_api.broker.delta_models.broker_zips import ZIPS_COLUMNS, zips_sql_string
 from usaspending_api.common.etl.spark import extract_db_data_frame, get_partition_bounds_sql, load_delta_table
 from usaspending_api.common.helpers.spark_helpers import (
     configure_spark_session,
@@ -20,274 +10,7 @@ from usaspending_api.common.helpers.spark_helpers import (
     get_jvm_logger,
 )
 from usaspending_api.config import CONFIG
-from usaspending_api.recipient.delta_models import (
-    RECIPIENT_LOOKUP_COLUMNS,
-    recipient_lookup_create_sql_string,
-    recipient_profile_create_sql_string,
-    RECIPIENT_PROFILE_DELTA_COLUMNS,
-    SAM_RECIPIENT_COLUMNS,
-    sam_recipient_create_sql_string,
-)
-from usaspending_api.search.models import TransactionSearch, AwardSearch
-from usaspending_api.transactions.delta_models import (
-    DETACHED_AWARD_PROCUREMENT_DELTA_COLUMNS,
-    detached_award_procurement_create_sql_string,
-    TRANSACTION_FABS_VIEW_COLUMNS,
-    transaction_fabs_sql_string,
-    TRANSACTION_FPDS_VIEW_COLUMNS,
-    transaction_fpds_sql_string,
-    TRANSACTION_NORMALIZED_COLUMNS,
-    transaction_normalized_sql_string,
-    TRANSACTION_SEARCH_POSTGRES_COLUMNS,
-    transaction_search_create_sql_string,
-    PUBLISHED_FABS_COLUMNS,
-    published_fabs_create_sql_string,
-)
-from usaspending_api.transactions.models import SourceAssistanceTransaction
-from usaspending_api.transactions.models import SourceProcurementTransaction
-from usaspending_api.search.delta_models.award_search import award_search_create_sql_string, AWARD_SEARCH_COLUMNS
-
-from usaspending_api.recipient.models import DUNS, RecipientLookup, RecipientProfile
-from usaspending_api.awards.models import (
-    Award,
-    FinancialAccountsByAwards,
-    TransactionFABS,
-    TransactionFPDS,
-    TransactionNormalized,
-)
-
-
-TABLE_SPEC = {
-    "awards": {
-        "model": Award,
-        "is_from_broker": False,
-        "source_table": "vw_awards",
-        "source_database": "rpt",
-        "destination_database": "raw",
-        "swap_schema": None,
-        "partition_column": "id",
-        "partition_column_type": "numeric",
-        "is_partition_column_unique": True,
-        "delta_table_create_sql": awards_sql_string,
-        "source_schema": None,
-        "custom_schema": "",
-        "column_names": list(AWARDS_COLUMNS),
-        "tsvectors": None,
-    },
-    "detached_award_procurement": {
-        "model": SourceProcurementTransaction,
-        "is_from_broker": False,
-        "source_table": "source_procurement_transaction",
-        "source_database": "raw",
-        "destination_database": "raw",
-        "swap_schema": None,
-        "partition_column": "detached_award_procurement_id",
-        "partition_column_type": "numeric",
-        "is_partition_column_unique": True,
-        "delta_table_create_sql": detached_award_procurement_create_sql_string,
-        "source_schema": None,
-        "custom_schema": "",
-        "column_names": list(DETACHED_AWARD_PROCUREMENT_DELTA_COLUMNS),
-        "tsvectors": None,
-    },
-    "financial_accounts_by_awards": {
-        "model": FinancialAccountsByAwards,
-        "is_from_broker": False,
-        "source_table": "financial_accounts_by_awards",
-        "source_database": "public",
-        "destination_database": "raw",
-        "swap_schema": None,
-        "partition_column": "financial_accounts_by_awards_id",
-        "partition_column_type": "numeric",
-        "is_partition_column_unique": True,
-        "delta_table_create_sql": financial_accounts_by_awards_sql_string,
-        "source_schema": None,
-        "custom_schema": "award_id LONG",
-        "column_names": list(FINANCIAL_ACCOUNTS_BY_AWARDS_COLUMNS),
-        "tsvectors": None,
-    },
-    "transaction_fabs": {
-        "model": TransactionFABS,
-        "is_from_broker": False,
-        "source_table": "vw_transaction_fabs",
-        "source_database": "int",
-        "destination_database": "raw",
-        "swap_schema": None,
-        "partition_column": "transaction_id",
-        "partition_column_type": "numeric",
-        "is_partition_column_unique": True,
-        "delta_table_create_sql": transaction_fabs_sql_string,
-        "source_schema": None,
-        "custom_schema": "",
-        "column_names": TRANSACTION_FABS_VIEW_COLUMNS,
-        "tsvectors": None,
-    },
-    "published_fabs": {
-        "model": SourceAssistanceTransaction,
-        "is_from_broker": False,
-        "source_table": "source_assistance_transaction",
-        "source_database": "raw",
-        "destination_database": "raw",
-        "swap_schema": None,
-        "partition_column": "published_fabs_id",
-        "partition_column_type": "numeric",
-        "is_partition_column_unique": True,
-        "delta_table_create_sql": published_fabs_create_sql_string,
-        "source_schema": None,
-        "custom_schema": "",
-        "column_names": list(PUBLISHED_FABS_COLUMNS),
-        "tsvectors": None,
-    },
-    "transaction_fpds": {
-        "model": TransactionFPDS,
-        "is_from_broker": False,
-        "source_table": "vw_transaction_fpds",
-        "source_database": "int",
-        "destination_database": "raw",
-        "swap_schema": None,
-        "partition_column": "transaction_id",
-        "partition_column_type": "numeric",
-        "is_partition_column_unique": True,
-        "delta_table_create_sql": transaction_fpds_sql_string,
-        "source_schema": None,
-        "custom_schema": "",
-        "column_names": TRANSACTION_FPDS_VIEW_COLUMNS,
-        "tsvectors": None,
-    },
-    "transaction_normalized": {
-        "model": TransactionNormalized,
-        "is_from_broker": False,
-        "source_table": "vw_transaction_normalized",
-        "source_database": "int",
-        "destination_database": "raw",
-        "swap_schema": None,
-        "partition_column": "id",
-        "partition_column_type": "numeric",
-        "is_partition_column_unique": True,
-        "delta_table_create_sql": transaction_normalized_sql_string,
-        "source_schema": None,
-        "custom_schema": "",
-        "column_names": list(TRANSACTION_NORMALIZED_COLUMNS),
-        "tsvectors": None,
-    },
-    # Tables loaded in from the Broker
-    "subaward": {
-        "model": None,
-        "is_from_broker": True,
-        "source_table": "subaward",
-        "source_database": None,
-        "destination_database": "raw",
-        "swap_schema": None,
-        "partition_column": "id",
-        "partition_column_type": "numeric",
-        "is_partition_column_unique": True,
-        "delta_table_create_sql": broker_subawards_sql_string,
-        "source_schema": None,
-        "custom_schema": "",
-        "column_names": list(BROKER_SUBAWARDS_COLUMNS),
-        "tsvectors": None,
-    },
-    "zips": {
-        "model": None,
-        "is_from_broker": True,
-        "source_table": "zips",
-        "source_database": None,
-        "destination_database": "raw",
-        "swap_schema": None,
-        "partition_column": "zips_id",
-        "partition_column_type": "numeric",
-        "is_partition_column_unique": True,
-        "delta_table_create_sql": zips_sql_string,
-        "source_schema": None,
-        "custom_schema": "",
-        "column_names": list(ZIPS_COLUMNS),
-        "tsvectors": None,
-    },
-    # Additional definitions for use in testing;
-    # These are copies of Views / Materialized Views / Tables from Postgres to Spark to aid in
-    # data comparison between current Postgres data and the data transformed via Spark.
-    "award_search_testing": {
-        "model": AwardSearch,
-        "is_from_broker": False,
-        "source_table": "award_search",
-        "source_database": None,
-        "destination_database": "rpt",
-        "swap_schema": None,
-        "partition_column": "award_id",
-        "partition_column_type": "numeric",
-        "is_partition_column_unique": True,
-        "delta_table_create_sql": award_search_create_sql_string,
-        "source_schema": None,
-        "custom_schema": "total_covid_outlay NUMERIC(23,2), total_covid_obligation NUMERIC(23,2), recipient_hash "
-        "STRING, federal_accounts STRING, cfdas ARRAY<STRING>, tas_components ARRAY<STRING>",
-        "column_names": list(AWARD_SEARCH_COLUMNS),
-        "tsvectors": None,
-    },
-    "recipient_lookup_testing": {
-        "model": RecipientLookup,
-        "is_from_broker": False,
-        "source_table": "recipient_lookup",
-        "source_database": "rpt",
-        "destination_database": "raw",
-        "swap_schema": None,
-        "partition_column": "id",
-        "partition_column_type": "numeric",
-        "is_partition_column_unique": True,
-        "delta_table_create_sql": recipient_lookup_create_sql_string,
-        "source_schema": None,
-        "custom_schema": "recipient_hash STRING",
-        "column_names": list(RECIPIENT_LOOKUP_COLUMNS),
-        "tsvectors": None,
-    },
-    "recipient_profile_testing": {
-        "model": RecipientProfile,
-        "is_from_broker": False,
-        "source_table": "recipient_profile",
-        "source_database": "rpt",
-        "destination_database": "raw",
-        "swap_schema": None,
-        "partition_column": "id",
-        "partition_column_type": "numeric",
-        "delta_table_create_sql": recipient_profile_create_sql_string,
-        "is_partition_column_unique": True,
-        "source_schema": None,
-        "custom_schema": "recipient_hash STRING",
-        "column_names": list(RECIPIENT_PROFILE_DELTA_COLUMNS),
-        "tsvectors": None,
-    },
-    "sam_recipient_testing": {
-        "model": DUNS,
-        "is_from_broker": False,
-        "source_table": "duns",
-        "source_database": "int",
-        "destination_database": "raw",
-        "swap_schema": None,
-        "partition_column": None,
-        "partition_column_type": None,
-        "is_partition_column_unique": False,
-        "delta_table_create_sql": sam_recipient_create_sql_string,
-        "source_schema": None,
-        "custom_schema": "broker_duns_id STRING, business_types_codes ARRAY<STRING>",
-        "column_names": list(SAM_RECIPIENT_COLUMNS),
-        "tsvectors": None,
-    },
-    "transaction_search_testing": {
-        "model": TransactionSearch,
-        "is_from_broker": False,
-        "source_table": "transaction_search",
-        "source_database": None,
-        "destination_database": "test",
-        "swap_schema": None,
-        "partition_column": "transaction_id",
-        "partition_column_type": "numeric",
-        "is_partition_column_unique": True,
-        "delta_table_create_sql": transaction_search_create_sql_string,
-        "source_schema": None,
-        "custom_schema": "recipient_hash STRING, federal_accounts STRING, parent_recipient_hash STRING",
-        "column_names": list(TRANSACTION_SEARCH_POSTGRES_COLUMNS),
-        "tsvectors": None,
-    },
-}
+from usaspending_api.etl.management.helpers.table_specifications import POSTGRES_GENERATED_TABLE_SPEC as TABLE_SPEC
 
 SPARK_PARTITION_ROWS = CONFIG.SPARK_PARTITION_ROWS
