+++ usaspending_api/etl/management/helpers/table_specifications.py
@@ -1,0 +1,518 @@
+from usaspending_api.awards.delta_models import (
+    BROKER_SUBAWARDS_COLUMNS,
+    broker_subawards_sql_string,
+    AWARDS_COLUMNS,
+    awards_sql_string,
+    FINANCIAL_ACCOUNTS_BY_AWARDS_COLUMNS,
+    financial_accounts_by_awards_sql_string,
+)
+from usaspending_api.awards.models import (
+    Award,
+    FinancialAccountsByAwards,
+    TransactionFABS,
+    TransactionFPDS,
+    TransactionNormalized,
+)
+from usaspending_api.broker.delta_models.broker_zips import ZIPS_COLUMNS, zips_sql_string
+from usaspending_api.recipient.delta_models import (
+    RECIPIENT_LOOKUP_COLUMNS,
+    recipient_lookup_create_sql_string,
+    recipient_lookup_load_sql_string_list,
+    RECIPIENT_LOOKUP_POSTGRES_COLUMNS,
+    recipient_profile_create_sql_string,
+    RECIPIENT_PROFILE_DELTA_COLUMNS,
+    recipient_profile_load_sql_strings,
+    RECIPIENT_PROFILE_POSTGRES_COLUMNS,
+    rpt_recipient_lookup_create_sql_string,
+    RPT_RECIPIENT_LOOKUP_DELTA_COLUMNS,
+    RPT_RECIPIENT_PROFILE_DELTA_COLUMNS,
+    SAM_RECIPIENT_COLUMNS,
+    sam_recipient_create_sql_string,
+    sam_recipient_load_sql_string,
+    SAM_RECIPIENT_POSTGRES_COLUMNS,
+)
+from usaspending_api.recipient.models import DUNS, RecipientLookup, RecipientProfile
+from usaspending_api.search.delta_models.award_search import (
+    AWARD_SEARCH_COLUMNS,
+    award_search_create_sql_string,
+    award_search_incremental_delete_temp_schema,
+    award_search_incremental_load_sql_string,
+    award_search_overwrite_load_sql_string,
+    AWARD_SEARCH_POSTGRES_COLUMNS,
+)
+from usaspending_api.search.delta_models.subaward_search import (
+    SUBAWARD_SEARCH_COLUMNS,
+    subaward_search_create_sql_string,
+    subaward_search_load_sql_string,
+    SUBAWARD_SEARCH_POSTGRES_COLUMNS,
+    SUBAWARD_SEARCH_POSTGRES_VECTORS,
+)
+from usaspending_api.search.models import AwardSearch, SubawardSearch, SummaryStateView, TransactionSearch
+from usaspending_api.transactions.delta_models import (
+    detached_award_procurement_create_sql_string,
+    DETACHED_AWARD_PROCUREMENT_DELTA_COLUMNS,
+    PUBLISHED_FABS_COLUMNS,
+    published_fabs_create_sql_string,
+    SUMMARY_STATE_VIEW_COLUMNS,
+    summary_state_view_create_sql_string,
+    summary_state_view_load_sql_string,
+    SUMMARY_STATE_VIEW_POSTGRES_COLUMNS,
+    TRANSACTION_CURRENT_CD_LOOKUP_COLUMNS,
+    transaction_current_cd_lookup_create_sql_string,
+    transaction_current_cd_lookup_load_sql_string,
+    transaction_fabs_sql_string,
+    TRANSACTION_FABS_VIEW_COLUMNS,
+    transaction_fpds_sql_string,
+    TRANSACTION_FPDS_VIEW_COLUMNS,
+    TRANSACTION_NORMALIZED_COLUMNS,
+    transaction_normalized_sql_string,
+    transaction_search_create_sql_string,
+    transaction_search_load_sql_string,
+    TRANSACTION_SEARCH_POSTGRES_COLUMNS,
+    TRANSACTION_SEARCH_POSTGRES_GOLD_COLUMNS,
+)
+from usaspending_api.transactions.models import SourceAssistanceTransaction
+from usaspending_api.transactions.models import SourceProcurementTransaction
+
+
+DATABRICKS_GENERATED_TABLE_SPEC = {
+    "award_search": {
+        "model": AwardSearch,
+        "is_from_broker": False,
+        "source_query": award_search_overwrite_load_sql_string,
+        "source_query_incremental": award_search_incremental_load_sql_string,
+        "source_database": None,
+        "source_table": None,
+        "destination_database": "rpt",
+        "swap_table": "award_search",
+        "swap_schema": "rpt",
+        "unique_identifiers": ["award_id"],
+        "delta_table_create_sql": award_search_create_sql_string,
+        "source_schema": AWARD_SEARCH_POSTGRES_COLUMNS,
+        "custom_schema": "recipient_hash STRING, federal_accounts STRING, cfdas ARRAY<STRING>,"
+        " tas_components ARRAY<STRING>",
+        "column_names": list(AWARD_SEARCH_COLUMNS),
+        "postgres_seq_name": None,
+        "tsvectors": None,
+        "postgres_partition_spec": None,
+        # Incremental Updates
+        "source_query_incremental": award_search_incremental_load_sql_string,
+        "incremental_delete_temp_schema": award_search_incremental_delete_temp_schema,
+        "delta_table_load_version_key": "award_search",
+    },
+    "recipient_lookup": {
+        "model": RecipientLookup,
+        "is_from_broker": False,
+        "source_query": recipient_lookup_load_sql_string_list,
+        "source_database": None,
+        "source_table": None,
+        "destination_database": "rpt",
+        "swap_table": "recipient_lookup",
+        "swap_schema": "rpt",
+        "unique_identifiers": ["recipient_hash"],
+        "delta_table_create_sql": rpt_recipient_lookup_create_sql_string,
+        "source_schema": RECIPIENT_LOOKUP_POSTGRES_COLUMNS,
+        "custom_schema": "recipient_hash STRING",
+        "column_names": list(RPT_RECIPIENT_LOOKUP_DELTA_COLUMNS),
+        "postgres_seq_name": "recipient_lookup_id_seq",
+        "tsvectors": None,
+        "postgres_partition_spec": None,
+        # Incremental Updates
+        "source_query_incremental": None,
+        "incremental_delete_temp_schema": None,
+        "delta_table_load_version_key": None,
+    },
+    "recipient_profile": {
+        "model": RecipientProfile,
+        "is_from_broker": False,
+        "source_query": recipient_profile_load_sql_strings,
+        "source_database": None,
+        "source_table": None,
+        "destination_database": "rpt",
+        "swap_table": "recipient_profile",
+        "swap_schema": "rpt",
+        "unique_identifiers": ["recipient_hash", "recipient_level"],
+        "delta_table_create_sql": recipient_profile_create_sql_string,
+        "source_schema": RECIPIENT_PROFILE_POSTGRES_COLUMNS,
+        "custom_schema": "recipient_hash STRING",
+        "column_names": list(RPT_RECIPIENT_PROFILE_DELTA_COLUMNS),
+        "postgres_seq_name": "recipient_profile_id_seq",
+        "tsvectors": None,
+        "postgres_partition_spec": None,
+        # Incremental Updates
+        "source_query_incremental": None,
+        "incremental_delete_temp_schema": None,
+        "delta_table_load_version_key": None,
+    },
+    "summary_state_view": {
+        "model": SummaryStateView,
+        "is_from_broker": False,
+        "source_query": summary_state_view_load_sql_string,
+        "source_database": None,
+        "source_table": None,
+        "destination_database": "rpt",
+        "swap_table": "summary_state_view",
+        "swap_schema": "rpt",
+        "unique_identifiers": ["duh"],
+        "delta_table_create_sql": summary_state_view_create_sql_string,
+        "source_schema": SUMMARY_STATE_VIEW_POSTGRES_COLUMNS,
+        "custom_schema": "duh STRING",
+        "column_names": list(SUMMARY_STATE_VIEW_COLUMNS),
+        "postgres_seq_name": None,
+        "tsvectors": None,
+        "postgres_partition_spec": None,
+        # Incremental Updates
+        "source_query_incremental": None,
+        "incremental_delete_temp_schema": None,
+        "delta_table_load_version_key": None,
+    },
+    "sam_recipient": {
+        "model": None,
+        "is_from_broker": True,
+        "source_query": sam_recipient_load_sql_string,
+        "source_database": None,
+        "source_table": None,
+        "destination_database": "int",
+        "swap_table": "duns",
+        "swap_schema": "int",
+        "unique_identifiers": ["broker_duns_id"],
+        "delta_table_create_sql": sam_recipient_create_sql_string,
+        "source_schema": SAM_RECIPIENT_POSTGRES_COLUMNS,
+        "custom_schema": None,
+        "column_names": list(SAM_RECIPIENT_COLUMNS),
+        "postgres_seq_name": None,
+        "tsvectors": None,
+        "postgres_partition_spec": None,
+        # Incremental Updates
+        "source_query_incremental": None,
+        "incremental_delete_temp_schema": None,
+        "delta_table_load_version_key": None,
+    },
+    "transaction_search": {
+        "model": TransactionSearch,
+        "is_from_broker": False,
+        "source_query": transaction_search_load_sql_string,
+        "source_database": None,
+        "source_table": None,
+        "destination_database": "rpt",
+        "swap_table": "transaction_search",
+        "swap_schema": "rpt",
+        "unique_identifiers": ["transaction_id"],
+        "delta_table_create_sql": transaction_search_create_sql_string,
+        "source_schema": TRANSACTION_SEARCH_POSTGRES_COLUMNS,
+        "custom_schema": "recipient_hash STRING, federal_accounts STRING, parent_recipient_hash STRING",
+        "column_names": list(TRANSACTION_SEARCH_POSTGRES_COLUMNS),
+        "postgres_seq_name": None,
+        "tsvectors": None,
+        "postgres_partition_spec": None,
+        # Incremental Updates
+        "source_query_incremental": None,
+        "incremental_delete_temp_schema": None,
+        "delta_table_load_version_key": None,
+    },
+    "transaction_search_gold": {
+        "model": TransactionSearch,
+        "is_from_broker": False,
+        "source_query": transaction_search_load_sql_string,
+        "source_database": None,
+        "source_table": None,
+        "destination_database": "rpt",
+        "swap_table": "transaction_search",
+        "swap_schema": "rpt",
+        "unique_identifiers": ["transaction_id"],
+        "delta_table_create_sql": transaction_search_create_sql_string,
+        "source_schema": TRANSACTION_SEARCH_POSTGRES_GOLD_COLUMNS,
+        "custom_schema": "recipient_hash STRING, federal_accounts STRING, parent_recipient_hash STRING",
+        "column_names": list(TRANSACTION_SEARCH_POSTGRES_GOLD_COLUMNS),
+        "postgres_seq_name": None,
+        "tsvectors": None,
+        "postgres_partition_spec": {
+            "partition_keys": ["is_fpds"],
+            "partitioning_form": "LIST",
+            "partitions": [
+                {"table_suffix": "_fpds", "partitioning_clause": "FOR VALUES IN (TRUE)"},
+                {"table_suffix": "_fabs", "partitioning_clause": "FOR VALUES IN (FALSE)"},
+            ],
+        },
+        # Incremental Updates
+        "source_query_incremental": None,
+        "incremental_delete_temp_schema": None,
+        "delta_table_load_version_key": None,
+    },
+    "transaction_current_cd_lookup": {
+        "model": None,
+        "is_from_broker": False,
+        "source_query": transaction_current_cd_lookup_load_sql_string,
+        "source_database": None,
+        "source_table": None,
+        "destination_database": "int",
+        "swap_table": "transaction_current_cd_lookup",
+        "swap_schema": "int",
+        "unique_identifiers": ["transaction_id"],
+        "delta_table_create_sql": transaction_current_cd_lookup_create_sql_string,
+        "source_schema": TRANSACTION_CURRENT_CD_LOOKUP_COLUMNS,
+        "custom_schema": "",
+        "column_names": list(TRANSACTION_CURRENT_CD_LOOKUP_COLUMNS),
+        "postgres_seq_name": None,
+        "tsvectors": None,
+        "postgres_partition_spec": None,
+        # Incremental Updates
+        "source_query_incremental": None,
+        "incremental_delete_temp_schema": None,
+        "delta_table_load_version_key": None,
+    },
+    "subaward_search": {
+        "model": SubawardSearch,
+        "is_from_broker": False,
+        "source_query": subaward_search_load_sql_string,
+        "source_database": None,
+        "source_table": None,
+        "destination_database": "rpt",
+        "swap_table": "subaward_search",
+        "swap_schema": "rpt",
+        "unique_identifiers": ["broker_subaward_id"],
+        "delta_table_create_sql": subaward_search_create_sql_string,
+        "source_schema": SUBAWARD_SEARCH_POSTGRES_COLUMNS,
+        "custom_schema": "treasury_account_identifiers ARRAY<INTEGER>",
+        "column_names": list(SUBAWARD_SEARCH_COLUMNS),
+        "postgres_seq_name": None,
+        "tsvectors": SUBAWARD_SEARCH_POSTGRES_VECTORS,
+        "postgres_partition_spec": None,
+        # Incremental Updates
+        "source_query_incremental": None,
+        "incremental_delete_temp_schema": None,
+        "delta_table_load_version_key": None,
+    },
+}
+
+POSTGRES_GENERATED_TABLE_SPEC = {
+    "awards": {
+        "model": Award,
+        "is_from_broker": False,
+        "source_table": "vw_awards",
+        "source_database": "rpt",
+        "destination_database": "raw",
+        "swap_schema": None,
+        "partition_column": "id",
+        "partition_column_type": "numeric",
+        "is_partition_column_unique": True,
+        "delta_table_create_sql": awards_sql_string,
+        "source_schema": None,
+        "custom_schema": "",
+        "column_names": list(AWARDS_COLUMNS),
+        "tsvectors": None,
+    },
+    "detached_award_procurement": {
+        "model": SourceProcurementTransaction,
+        "is_from_broker": False,
+        "source_table": "source_procurement_transaction",
+        "source_database": "raw",
+        "destination_database": "raw",
+        "swap_schema": None,
+        "partition_column": "detached_award_procurement_id",
+        "partition_column_type": "numeric",
+        "is_partition_column_unique": True,
+        "delta_table_create_sql": detached_award_procurement_create_sql_string,
+        "source_schema": None,
+        "custom_schema": "",
+        "column_names": list(DETACHED_AWARD_PROCUREMENT_DELTA_COLUMNS),
+        "tsvectors": None,
+    },
+    "financial_accounts_by_awards": {
+        "model": FinancialAccountsByAwards,
+        "is_from_broker": False,
+        "source_table": "financial_accounts_by_awards",
+        "source_database": "public",
+        "destination_database": "raw",
+        "swap_schema": None,
+        "partition_column": "financial_accounts_by_awards_id",
+        "partition_column_type": "numeric",
+        "is_partition_column_unique": True,
+        "delta_table_create_sql": financial_accounts_by_awards_sql_string,
+        "source_schema": None,
+        "custom_schema": "award_id LONG",
+        "column_names": list(FINANCIAL_ACCOUNTS_BY_AWARDS_COLUMNS),
+        "tsvectors": None,
+    },
+    "transaction_fabs": {
+        "model": TransactionFABS,
+        "is_from_broker": False,
+        "source_table": "vw_transaction_fabs",
+        "source_database": "int",
+        "destination_database": "raw",
+        "swap_schema": None,
+        "partition_column": "transaction_id",
+        "partition_column_type": "numeric",
+        "is_partition_column_unique": True,
+        "delta_table_create_sql": transaction_fabs_sql_string,
+        "source_schema": None,
+        "custom_schema": "",
+        "column_names": TRANSACTION_FABS_VIEW_COLUMNS,
+        "tsvectors": None,
+    },
+    "published_fabs": {
+        "model": SourceAssistanceTransaction,
+        "is_from_broker": False,
+        "source_table": "source_assistance_transaction",
+        "source_database": "raw",
+        "destination_database": "raw",
+        "swap_schema": None,
+        "partition_column": "published_fabs_id",
+        "partition_column_type": "numeric",
+        "is_partition_column_unique": True,
+        "delta_table_create_sql": published_fabs_create_sql_string,
+        "source_schema": None,
+        "custom_schema": "",
+        "column_names": list(PUBLISHED_FABS_COLUMNS),
+        "tsvectors": None,
+    },
+    "transaction_fpds": {
+        "model": TransactionFPDS,
+        "is_from_broker": False,
+        "source_table": "vw_transaction_fpds",
+        "source_database": "int",
+        "destination_database": "raw",
+        "swap_schema": None,
+        "partition_column": "transaction_id",
+        "partition_column_type": "numeric",
+        "is_partition_column_unique": True,
+        "delta_table_create_sql": transaction_fpds_sql_string,
+        "source_schema": None,
+        "custom_schema": "",
+        "column_names": TRANSACTION_FPDS_VIEW_COLUMNS,
+        "tsvectors": None,
+    },
+    "transaction_normalized": {
+        "model": TransactionNormalized,
+        "is_from_broker": False,
+        "source_table": "vw_transaction_normalized",
+        "source_database": "int",
+        "destination_database": "raw",
+        "swap_schema": None,
+        "partition_column": "id",
+        "partition_column_type": "numeric",
+        "is_partition_column_unique": True,
+        "delta_table_create_sql": transaction_normalized_sql_string,
+        "source_schema": None,
+        "custom_schema": "",
+        "column_names": list(TRANSACTION_NORMALIZED_COLUMNS),
+        "tsvectors": None,
+    },
+    # Tables loaded in from the Broker
+    "subaward": {
+        "model": None,
+        "is_from_broker": True,
+        "source_table": "subaward",
+        "source_database": None,
+        "destination_database": "raw",
+        "swap_schema": None,
+        "partition_column": "id",
+        "partition_column_type": "numeric",
+        "is_partition_column_unique": True,
+        "delta_table_create_sql": broker_subawards_sql_string,
+        "source_schema": None,
+        "custom_schema": "",
+        "column_names": list(BROKER_SUBAWARDS_COLUMNS),
+        "tsvectors": None,
+    },
+    "zips": {
+        "model": None,
+        "is_from_broker": True,
+        "source_table": "zips",
+        "source_database": None,
+        "destination_database": "raw",
+        "swap_schema": None,
+        "partition_column": "zips_id",
+        "partition_column_type": "numeric",
+        "is_partition_column_unique": True,
+        "delta_table_create_sql": zips_sql_string,
+        "source_schema": None,
+        "custom_schema": "",
+        "column_names": list(ZIPS_COLUMNS),
+        "tsvectors": None,
+    },
+    # Additional definitions for use in testing;
+    # These are copies of Views / Materialized Views / Tables from Postgres to Spark to aid in
+    # data comparison between current Postgres data and the data transformed via Spark.
+    "award_search_testing": {
+        "model": AwardSearch,
+        "is_from_broker": False,
+        "source_table": "award_search",
+        "source_database": None,
+        "destination_database": "rpt",
+        "swap_schema": None,
+        "partition_column": "award_id",
+        "partition_column_type": "numeric",
+        "is_partition_column_unique": True,
+        "delta_table_create_sql": award_search_create_sql_string,
+        "source_schema": None,
+        "custom_schema": "total_covid_outlay NUMERIC(23,2), total_covid_obligation NUMERIC(23,2), recipient_hash "
+        "STRING, federal_accounts STRING, cfdas ARRAY<STRING>, tas_components ARRAY<STRING>",
+        "column_names": list(AWARD_SEARCH_COLUMNS),
+        "tsvectors": None,
+    },
+    "recipient_lookup_testing": {
+        "model": RecipientLookup,
+        "is_from_broker": False,
+        "source_table": "recipient_lookup",
+        "source_database": "rpt",
+        "destination_database": "raw",
+        "swap_schema": None,
+        "partition_column": "id",
+        "partition_column_type": "numeric",
+        "is_partition_column_unique": True,
+        "delta_table_create_sql": recipient_lookup_create_sql_string,
+        "source_schema": None,
+        "custom_schema": "recipient_hash STRING",
+        "column_names": list(RECIPIENT_LOOKUP_COLUMNS),
+        "tsvectors": None,
+    },
+    "recipient_profile_testing": {
+        "model": RecipientProfile,
+        "is_from_broker": False,
+        "source_table": "recipient_profile",
+        "source_database": "rpt",
+        "destination_database": "raw",
+        "swap_schema": None,
+        "partition_column": "id",
+        "partition_column_type": "numeric",
+        "delta_table_create_sql": recipient_profile_create_sql_string,
+        "is_partition_column_unique": True,
+        "source_schema": None,
+        "custom_schema": "recipient_hash STRING",
+        "column_names": list(RECIPIENT_PROFILE_DELTA_COLUMNS),
+        "tsvectors": None,
+    },
+    "sam_recipient_testing": {
+        "model": DUNS,
+        "is_from_broker": False,
+        "source_table": "duns",
+        "source_database": "int",
+        "destination_database": "raw",
+        "swap_schema": None,
+        "partition_column": None,
+        "partition_column_type": None,
+        "is_partition_column_unique": False,
+        "delta_table_create_sql": sam_recipient_create_sql_string,
+        "source_schema": None,
+        "custom_schema": "broker_duns_id STRING, business_types_codes ARRAY<STRING>",
+        "column_names": list(SAM_RECIPIENT_COLUMNS),
+        "tsvectors": None,
+    },
+    "transaction_search_testing": {
+        "model": TransactionSearch,
+        "is_from_broker": False,
+        "source_table": "transaction_search",
+        "source_database": None,
+        "destination_database": "test",
+        "swap_schema": None,
+        "partition_column": "transaction_id",
+        "partition_column_type": "numeric",
+        "is_partition_column_unique": True,
+        "delta_table_create_sql": transaction_search_create_sql_string,
+        "source_schema": None,
+        "custom_schema": "recipient_hash STRING, federal_accounts STRING, parent_recipient_hash STRING",
+        "column_names": list(TRANSACTION_SEARCH_POSTGRES_COLUMNS),
+        "tsvectors": None,
+    },
+}
